feat: add getDockerTag helper
