+++ lib/web-service/index.ts
@@ -1,6 +1,6 @@
 export * from "./annotation-util";
+export * from "./env-util";
 export * from "./nginx-util";
 export * from "./resque-web";
-export * from "./stage-util";
 export * from "./web-service-props";
 export * from "./web-service";
